Click on instrument names in sequencer to open controller.
This reveals a couple issues. First, synth controllers don't unhook themselves
when closed. This is missing in seesaw.bind, so there's some work to do
there. Second, on OSX at least, the L&F for a toggle button doesn't show
its state when its not in the active window. So, when you show the synth
controller, the sequencer grid goes blank. We should probably just do a
custom row widget. Good times.