Now explicitly performing topological sort.
Rather than rely on a lexical binding that keeps track of all
ugens in a synthdef in the order they are created, we now take the root
of the ugen tree and do a topological sort.  This is one step
on the path towards more composable synths.

Topological sorting means that we start at the outer most leaf nodes
and work our way in such that no ugen is ordered before any of its
upstream dependencies.  Additionally, the sort algorithm will try
to place ugens along a linear branch in the tree one after another
so the audio data they are passing one to the next can reside in
cache for as long as possible.