Adding back the darn *ugens* and *constants* bindings.
We realized that in order to have side-effecting ugens within a synth, which
are not part of the processing tree, we have to use lexical variables
in order to keep track of the other ugens.  (e.g. detect-silence wouldn't work
because it wasn't in the topologically sorted tree of ugens.)

Bindings are back, and detect-silence and writing wavs to disk work again. Fixed.