Create sc.foundation-groups ns to create and represent foundation groups.
These groups are guaranteed to exist for all Overtone sessions. They contain no synths, and just represent a sensible empty structure to place your own synths. This structure is also used by other Overtone namespaces such as studio and scope. 

The core group is the foundation-default-group which is the default group for the group fn. In addition, there are two 'safe' groups which will not be affected by stop and friends.

Other changes:

 * stop no longer destroys all sub groups within the default group
 * there is a new 'user' group containing all user-oriented groups and synths
 * group ids are made available by the various foundation-*-group fns