Accessing buffer data at last!!!
Thanks for the hard work of Roger Allen, Markus Gustavsson, and Joel J., we've
now got direct access to buffer data from our internally booted SC server.

Now that this works I'm sure there is more work to be done to integrate some of
the other buffer communication mechanisms, but this is enough to at least start
using the scope again.