use clojure's builtin into-array fn
It looks like byte-seq-to-array is duplicating work that into-array can
do. This commit uses the builtin instead.

Note: I've tested this at the unit level, but as I can't seem to light
up the LEDs on my nanoKONTROL2 I haven't tested that it still works when
integrating with a real device.