Controlling non-ready nodes is now delayed until node is ready.
Created new protocol ISynthNodeStatus with two fn slots - node-status and node-block-until-ready which is now used to extend SynthNode, SynthGroup and Inst records.

Add loaded? slot in synth node records which contains a promise which is delivered when the node is loaded. This is used to implement node-block-until-ready. This allows for blocking of the current thread when not loaded - which is what now happens when a node modification fn is called via ensure-node-active!

Fixes a race condition when nodes were created and then immediately controlled/killed resulting in abnormal behaviour.