*Breaking change* - synth fn not expects a vector of node target/pos information rather than as 'special' initial args.
Instead of:

(foo :target my-b :position :tail :eggs 1 :beans 2)

You now need to write:

(foo [:tail my-b] :eggs 1 :beans 2)

Where the first element of the vector is the position and the second 
is the target.

Using the old style will now throw an exception explaining that way
is now deprecated.