Reserve control-bus 0 as a precautionary measure.
Many synthdefs with control-bus args use id 0 as the default value. If
these synths are triggered without passing a specific bus id as a param
then the best that can happen is that the synth behaves incorrectly as
it's reading from bus 0. However, the worst situation is when the synth
writes to bus 0 - in this case if bus 0 is already assigned for use, 
then the bus will be clobbered with data from the synth causing other
synths to potentially misbehave too.

By reserving bus 0, no allocated control-bus will have this id. 
Therefore, accidentally writing to this bus will not have any negative
impact on other correctly executing synths.