Wrap all calls to defrecord within defonce blocks.
When a defrecord form is re-evaluated, it creates a new Java class for
the record which is not .equal to the previous record, even if the
content of the defrecord form is unchanged. This can cause strange 
behaviour when instances of the old record fail to be recognised as 
instances of the new record. Guarding defrecord forms from re-evaluation
mitigates this issue.