Keep metronome consistent under multithreading
As it stood, the metronome could become inconsistent when multiple
threads are making changes to it at the same time, because it was using
atoms for its internal state, and atoms only provide isolated
consistency. Since changes sometimes need to be made to multiple state
items in a coordinated fashion, the metronome needs to instead use refs
and Clojure's software transactional memory features. I discovered this
problem because I am synchronizing a metronome to the BPM of DJ software
using MIDI clock pulses, while also allowing the user to beat jump, or
change features like beats-per-bar.