Fix broken version comparison in args/SC-ARG-INFO
The previous logic in the modified code compares the major and minor
portions of the supercollider version by converting the version string
to a floating point number and comparing it to a numeric constant. This
comparison breaks when upgrading to supercollider `3.10` because `(< 3.6
3.10)` is false.

This commit changes the comparison to use a sequence of integers in a
vector, such as `[3 10]`. This data structure can be compared using the
`compare` function resulting in semantics that match how semantic
versions are compared.

```
(neg? (compare [3 6] [3 5]))  ; => false
(neg? (compare [3 6] [3 6]))  ; => false
(neg? (compare [3 6] [3 9]))  ; => true
(neg? (compare [3 6] [3 10])) ; => true
```