Fix 21-arity version of clojure.lang.IFn
Besides the versions with a fixed number of arguments, there is one final one
that takes an array as its final argument (i.e. a Java varargs function). Take
that into account so we don't run into AbstractMethod error when it gets called.

Also deals with an issue using an unquoted `abs` in the stringed instrument, and
some minor namespace cleanup.