Mention `:force-ugen` in the colliding ugen docstrings
Some UGens collide with clojure.core functions, we have heuristics to figure out
if in a given context you want one or the other. If that fails, there is :force-ugen
as a hint. That is not really documented anywhere it seems. As a start mention it
in the generated docstrings for these functions.