Fix `defsynth-load`
We may want to defined a synth like below

```clojure
(defsynth-load directional
  (str "test-resources" "/directional.scsyndef"))
```

Which was throwing an exception because it didn't understand the
unevaluated `(str ...)` as a file path.

This fixes it.

The reason for using `defsynth-load` instead of `synth-load` is so we
can have a `defsynth`-like metadata attached to the var.